Don't hijack ActiveRecord::Base connection
The default of a sqlite in-memory database is somewhat presumptuous.
It was causing some trouble in the latest release where it would
attempt to connect to sqlite even though the it was already connected
to another database.

Rather than trying to get it to be more polite, I'm going to remove it
entirely. It's doing more harm than good if you have to worry about your
your ActiveRecord connection being hijacked depending on your test run
order.