Add automated multi-database tests
This commit automates testing against different database adapters.
Currently postgresql, mysql, mysql2 and sqlite3 are supported. The
adapters are configured in spec/config.default.yml. The rake tasks for
testing against each of the adapters are spec:postgresql, spec:mysql,
spec:mysql2 and spec:sqlite3. Additionally the spec task runs all these
tasks in a sequence.

PostgreSQL and MySQL should be configured for testing by running the
db:postgresql:create and db:mysql:create tasks.

This code is loosely based on the Active Record test suite.

A minor fix was required to make patching SQLite3Adapter work.
Previously ::establish_connection autoloaded SQLite3Adapter. The patch
opened the already-defined class and added two methods
(#temporary_tables and #temporary_table_exists?). Currently, a
connection is established in a before(:suite) block. It's called after
spec_helper.rb is run. Because of that the SQLite3Adapter patch in
spec_helper.rb became the first-time definition of the class. When the
original adapter was loaded it specified a different base class which
resulted in an exception reporting incompatible base-classes.
Loading all adapters, thus ensuring we patch classes that are already
loaded, may be a better idea but I decided to take a simpler approach
for now.

Last but not least CONTRIBUTING.md is added. It mentions, among other
things, how to set up and run the test suite before submitting a pull
request.